+++ packages/js-sdk/src/regex/detectors/eu/index.ts
@@ -1,0 +1,4 @@
+// EU-specific PII detectors
+import './iban'
+import './postal-code'
+import './vat-id'