+++ packages/js-sdk/src/regex/detectors/uk/index.ts
@@ -1,0 +1,5 @@
+// UK-specific PII detectors
+import './ni-number'
+import './nhs-number'
+import './postcode'
+import './passport'